refactor:이미지 수정 쿼리 커스텀훅 작성
