+++ src/components/modal/ProfileEditModal.tsx
@@ -1,13 +1,9 @@
 import Button from '@/components/common/Button';
 import { type FC, useRef, useCallback, useState } from 'react';
 import './profileEditModal.scss';
-import { patchProfileImage } from '@/api/auth/authAPI';
-import { useMutation } from '@tanstack/react-query';
-import axios from 'axios';
 import UserProfileImg from '@/components/common/UserProfileImg';
 import useAuth from '@/hooks/useAuth';
-import { useSetRecoilState } from 'recoil';
-import { userState } from '@/atom/atom';
+import usePatchProfileImgMutation from '@/queries/user/usePatchProfileImgMutation';
 
 interface Props {
   onClose: () => void;
@@ -17,32 +13,10 @@ interface Props {
 const ProfileEditModal: FC<Props> = ({ onClose, src }: Props) => {
   const inputRef = useRef<HTMLInputElement | null>(null);
   const { userId } = useAuth();
-  const setUser = useSetRecoilState(userState);
   const [previewSrc, setPreviewSrc] = useState(src);
   const [profileSrc, setProfileSrc] = useState<File | null>(null);
 
-  const { mutate: patchProfileImg } = useMutation(patchProfileImage, {
-    onSuccess: (res) => {
-      alert('수정 완료');
-      const profileUrl = res?.data?.profileUrl;
-      setUser((prev) => {
-        return prev ? { ...prev, profileUrl } : prev;
-      });
-      onClose();
-    },
-    onError: (err) => {
-      if (axios.isAxiosError(err)) {
-        const statusCode = err.response?.status;
-        const errorMessage = err.response?.data?.message;
-        if (statusCode === 400 && errorMessage === 'Max file size 2MB') {
-          alert('파일이 2MB를 초과하였습니다.');
-        }
-        if (statusCode === 400 && errorMessage === 'Invalid Values') {
-          alert('jpg/jpeg, png, gif 파일만 업로드 가능합니다.');
-        }
-      }
-    },
-  });
+  const { mutate: patchProfileImg } = usePatchProfileImgMutation(onClose);
 
   const onChangeImage = (event: React.ChangeEvent<HTMLInputElement>) => {
     if (!event.target.files) {
